requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
yfinance>=0.2.36
mplfinance>=0.12.10b0
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.2.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
requests>=2.31.0
alpaca-py>=0.20.0
